Refactorizada librería
Se han refactorizado los nombres de la biblioteca routingSQL para seguir las mejores prácticas. Esto incluye garantizar la consistencia en las convenciones de nomenclatura, usar nombres significativos y adherirse a las directrices de PEP 8.
