+++ ckanext/fototeca/lib/routing_sql/base.py
